+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fix preedit string disapear when press hotkey [#310](https://github.com/Riey/kime/issues/310)
 * Make character typing order strict `ㅏ + ㄱ = ㅏㄱ`
 * Add `TreatJongseongAsChoseong` addon
+* Add `FlexibleComposeOrder` addon [#318](https://github.com/Riey/kime/issues/318)
 
 ## 1.1.3